+++ common/src/main/java/com/nemonotfound/nemos/copper/interfaces/CopperBucketItemGetter.java
@@ -1,8 +1,0 @@
-package com.nemonotfound.nemos.copper.interfaces;
-
-import net.minecraft.world.item.Item;
-
-public interface CopperBucketItemGetter {
-
-    Item nemosCopper$getCopperBucket();
-}