+++ fabric/src/main/java/com/nemonotfound/nemos/copper/NemosCopperFabric.java
@@ -1,11 +1,11 @@
 package com.nemonotfound.nemos.copper;
 
+import com.nemonotfound.nemos.copper.core.dispenser.ModDispenseItemBehavior;
 import com.nemonotfound.nemos.copper.helper.CauldronInteractionHelper;
+import com.nemonotfound.nemos.copper.helper.ItemReplacementHelper;
 import com.nemonotfound.nemos.copper.item.ModItems;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.registry.FuelRegistryEvents;
-import net.minecraft.core.dispenser.ShearsDispenseItemBehavior;
-import net.minecraft.world.level.block.DispenserBlock;
 
 public class NemosCopperFabric implements ModInitializer {
     
@@ -15,7 +15,8 @@ public class NemosCopperFabric implements ModInitializer {
         FuelRegistryEvents.BUILD.register((builder, context) ->
                 builder.add(ModItems.COPPER_LAVA_BUCKET.get(), 20000));
 
-        DispenserBlock.registerBehavior(ModItems.COPPER_SHEARS.get().asItem(), new ShearsDispenseItemBehavior());
+        ModDispenseItemBehavior.bootstrap();
         CauldronInteractionHelper.addCopperBucketInteractions();
+        ItemReplacementHelper.addToItemReplacementMaps();
     }
 }