+++ neoforge/src/main/java/com/nemonotfound/nemos/copper/events/ModEvents.java
@@ -1,10 +1,9 @@
 package com.nemonotfound.nemos.copper.events;
 
 import com.nemonotfound.nemos.copper.Constants;
+import com.nemonotfound.nemos.copper.core.dispenser.ModDispenseItemBehavior;
 import com.nemonotfound.nemos.copper.helper.CauldronInteractionHelper;
-import com.nemonotfound.nemos.copper.item.ModItems;
-import net.minecraft.core.dispenser.ShearsDispenseItemBehavior;
-import net.minecraft.world.level.block.DispenserBlock;
+import com.nemonotfound.nemos.copper.helper.ItemReplacementHelper;
 import net.neoforged.bus.api.EventPriority;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
@@ -15,7 +14,8 @@ public class ModEvents {
 
     @SubscribeEvent(priority = EventPriority.LOWEST)
     private static void registerDispenseBehavior(final FMLCommonSetupEvent event) {
-        event.enqueueWork(() -> DispenserBlock.registerBehavior(ModItems.COPPER_SHEARS.get().asItem(), new ShearsDispenseItemBehavior()));
+        event.enqueueWork(ModDispenseItemBehavior::bootstrap);
         event.enqueueWork(CauldronInteractionHelper::addCopperBucketInteractions);
+        event.enqueueWork(ItemReplacementHelper::addToItemReplacementMaps);
     }
 }