+++ Changelog.md
@@ -1,4 +1,13 @@
-# Changelog v1.4
+# Changelog v1.5
 
 ## Additions
-- Added **Copper Shield**
+- Added copper bars
+- Added dependency
+  - Nemo's Tags
+
+## Changes
+- Internal changes & cleanup
+
+## Fixes
+- Copper bucket can't milk goats
+- Copper buckets can't be used in dispenser